Struct server_forge::rollback::RollbackManager

pub struct RollbackManager { /* private fields */ }

Manages the creation of snapshots and rollback operations.

Implementations

impl RollbackManager

pub fn new() -> Self

Creates a new RollbackManager instance.

pub fn create_snapshot(&self) -> Result<usize, Box<dyn Error>>

Creates a new snapshot and returns its ID.

Errors

Returns an error if the snapshot creation fails.

pub fn add_file_change( &self, snapshot_id: usize, file_path: &str ) -> Result<(), Box<dyn Error>>

Adds a file change to a specific snapshot.

Arguments

snapshot_id - The ID of the snapshot to add the file change to

file_path - The path of the changed file

Errors

Returns an error if reading the file fails or if the snapshot ID is invalid.

pub fn add_package_installed( &self, snapshot_id: usize, package: &str ) -> Result<(), Box<dyn Error>>

Adds an installed package to a specific snapshot.

Arguments

snapshot_id - The ID of the snapshot to add the package to

package - The name of the installed package

Errors

Returns an error if the snapshot ID is invalid.

pub fn commit_snapshot(&self, _snapshot_id: usize) -> Result<(), Box<dyn Error>>

Commits a snapshot, finalizing its state.

This method is a placeholder and currently does nothing. It could be expanded to compress the snapshot or write it to disk.

Arguments

_snapshot_id - The ID of the snapshot to commit

pub fn rollback_all(&self) -> Result<(), Box<dyn Error>>

Rolls back all changes made since the first snapshot.

Errors

Returns an error if any part of the rollback process fails.

pub fn rollback_to(&self, snapshot_id: usize) -> Result<(), Box<dyn Error>>

Rolls back changes to a specific snapshot.

Arguments

snapshot_id - The ID of the snapshot to roll back to

Errors

Returns an error if the snapshot ID is invalid or if any part of the rollback process fails.